Reduce ICLMetric mem consumption (#17)
We do a F.log_softmax(lm_logits, dim=-1) -> torch.gather() operation when updating ICLMetric.

This materializes a large tensor of logprobs. In the past this hasnt been a big deal, but when training a Pipeline parallel model we are much more memory constrained on rank holding the LMHead. Additionally, if we use a memory-efficient CE implementation like Liger's FusedLinearCrossEntropy at training time then we can actually use more memory at eval time when we need to emit the full block of logits.

We can ~half the memory required by the activations by avoiding materializing the logprobs tensor (so instead of logits and logprobs, we just materialize the logits). This can be done by pushing the log_softmax inside the loop and recognizing that the log_softmax -> gather operation is just negative cross entropy. Cross entropy implementations avoid materializing the full logprobs tensor under the hood.

This is one step towards lowering memory consumption.